FIX FT_CD AND FD_PWD:
Fixing CD and PWD, not logging precise changes because
this is the commit of interest on CD and PWD after WIP.

Fixed several typos.
Added function prototypes to .h file
Adding sources in Makefile

First test done on pwd
cd still needs to be thoroughly tested.

Changes to be committed:
	modified:   ../../Makefile
	modified:   ../../includes/minishell.h
	modified:   main_test_staf.c
	modified:   ms_cd.c
	modified:   ms_pwd.c